Fix to change the WorkitemKeySequence data type to BIGINT

+++ src/Microsoft.Health.Dicom.SqlServer/Features/Schema/Sql/Scripts/Sequences.sql
@@ -53,7 +53,7 @@ CREATE SEQUENCE dbo.PartitionKeySequence
     CACHE 10000
 
 CREATE SEQUENCE dbo.WorkitemKeySequence
-    AS INT
+    AS BIGINT
     START WITH 1
     INCREMENT BY 1
     MINVALUE 1
